+++ hl.vcxproj
@@ -290,6 +290,8 @@
     <ClCompile Include="src\code.c" />
     <ClCompile Include="src\debugger.c" />
     <ClCompile Include="src\jit.c" />
+    <ClCompile Include="src\jit_dump.c" />
+    <ClCompile Include="src\jit_emit.c" />
     <ClCompile Include="src\main.c" />
     <ClCompile Include="src\module.c" />
     <ClCompile Include="src\profile.c" />
@@ -298,6 +300,7 @@
     <ClInclude Include="src\hl.h" />
     <ClInclude Include="src\hlmodule.h" />
     <ClInclude Include="src\hlsystem.h" />
+    <ClInclude Include="src\jit.h" />
     <ClInclude Include="src\opcodes.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
