+++ hl.vcxproj.filters
@@ -4,14 +4,17 @@
     <ClCompile Include="src\main.c" />
     <ClCompile Include="src\code.c" />
     <ClCompile Include="src\module.c" />
-    <ClCompile Include="src\jit.c" />
     <ClCompile Include="src\debugger.c" />
     <ClCompile Include="src\profile.c" />
+    <ClCompile Include="src\jit_dump.c" />
+    <ClCompile Include="src\jit_emit.c" />
+    <ClCompile Include="src\jit.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\hlmodule.h" />
     <ClInclude Include="src\opcodes.h" />
     <ClInclude Include="src\hl.h" />
     <ClInclude Include="src\hlsystem.h" />
+    <ClInclude Include="src\jit.h" />
   </ItemGroup>
 </Project>
